rename: TableRowAccount
this component should rather be a story / should not exist
