+++ packages/ui-components/src/index.ts
@@ -60,7 +60,7 @@ export * from './Shared.styles';
 export * from './stateful';
 export * from './TableAccounts';
 export * from './TableCellAddress';
-export * from './TableRowsAccount';
+export * from './TableRowAccount';
 export * from './TextArea';
 export * from './TxDetails';
 export * from './TxSummary';
